Set up project/metadata; finish eigenvector_generators
Set up the barebones project structure, including adding metadata to the .toml files and a description to README.md. Also, finished the source code and documentation for src/SDiagonalizability/eigenvector_generators.jl. These internal functions (not part of the public API, or at least not intended to be for now) serve as helpers to the stuff in laplacian_spectra, where we validate Laplacian integrality, independent S-eigenspaces, etc. (Still need to add unit tests, though.)