Integrate codecov + static analysis; polish up README
Integrated code coverage with codecov into the GitHub Actions workflows, along with a JuliaFormatter check. Added tests with Aqua.jl and JET.jl for static analysis, and refactored the (very incomplete) tests/ folder. Updated compatibility in Project.toml to Julia 1.10+ for compatibility with JET, and changed the test matrices in the CI workflow accordingly. Updated README.md with a table of badges, plus the logo.