WIP: Improving type stability (analysis with JET.jl)
Added a Base.eltype override to the eigenvector generators in src/eigenvector_generation.jl. Made an inner constructor to the (public) SpectrumIntegralResult struct in src/laplacian_spectra.jl to filter out unintended input. Added an assert statement to the Base.getproperty override of the (private) _LaplacianSpectrum01Neg struct in the same file for compiler inference with JET, and replaced a recursive getproperty call with getfield. Converted matrix to dense float array in _extract_independent_cols for type stability/performance with LinearAlgebra.qr (and added whitespace).

Finally, tightened the bound on the number of JET reports to 20 (exclusive)! Still quite a few JET reports/warnings to go fix, but many of them are stale/artifacts that can remain.