Flesh out documentation + inline comments (#36)
This PR continues to add/update docstrings and inline comments throughout the codebase. Recall that the current status of our project so far is that everything is working except for the src/s_bandwidth.jl file tying together the whole codebase into our overarching algorithm. (Of course, unit tests still need to be implemented.)

With this in mind, we note that this PR completes docstrings (complete with doctests + examples where appopriate) and inline comments for all top-level objects in src/ (and src/factories/) except for in src/laplacian_spectra.jl and in src/basis_search.jl, both of which are around half done. We choose to merge already to minimize merge conflicts with concurrent work.

We also make a minor change to the README and the documentation homepage, updating the installation requirements (clarifying that a Julia runtime with version 1.10+ is needed) and lower the threshold for JET static analysis reports to <25. Some further design choices are refined as well.