Enforce if's over short circuits
This PR changes .JuliaFormatter.toml to include 'short_circuit_to_if = true', ensuring that all conditionals use if statements instead of one-liner short circuits. We then apply the formatting to the codebase, inadvertently also catching a bug in the quasi-orthogonal basis search ('a || b && c' did not reach 'c' if 'a' was true).