Made several changes; see extended message
First, the 'minimize_s_bandwidth' function is renamed to 's_bandwidth'.

Second, further results/useful checks from Johnston and Plosker (2025) are integrated into the core logic.

Third, the 's_bandwidth' function is updated to not run redundant checks on eigenspaces for a k-basis if the previous run failed for some other eigenspaces; in this way, it is no longer a wrapper around the recognition algorithm due to non-independent runs for each k.

Fourth, 'is_s_diagonalizable' now returns its own 'SDiagonalizabilityResult' struct, not just a somewhat semantically inappropriate 'SBandRecognitionResult'.

Finally, the 'Basic Use' section is completed in the README, with the Base.show methods for the result structs overriden.