Standardize formatting of 'S'
This PR standardizes the formatting of 'S' in docstrings, surrounding it with double backticks every time unless it is in reference to a parameter (in which case single backticks are used).

It also fleshes out the docstring for the '_assert_graph_has_defined_s_bandwidth' method in utils, although that is not yet complete.